app/requirements.txt
# requirements.txt for Kompetenzen-Tool

# HTTP requests and ICS parsing
requests>=2.0
icalendar>=4.0

# Database ORM + PostgreSQL driver
SQLAlchemy>=2.0
psycopg2-binary>=2.9

# Web UI
streamlit>=1.26

# Data processing
pandas>=1.5
numpy>=1.24

# YAML support
PyYAML>=5.4

# Environment variable loading
python-dotenv>=1.0
